fix(frontend): Refactor BaseMap and Context to set transitioning
The primary fix here is to set the transitioning of the map when
the view state changes. The eventual aim is to do away with helper
functions that are currently used to get the view state.
